fix released lock for subprocesses
