+++ tests/examples/orphan-subprocess/playground/slow-startup/poll-down
@@ -1,0 +1,1 @@
+0.01
