+++ tests/examples/orphan-subprocess/playground/slow-startup/ready
@@ -1,0 +1,2 @@
+#!/bin/bash
+exec test -e readyfile
