+++ tests/examples/orphan-subprocess/playground/slow-startup/run
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -eu
+# this service takes two seconds to become ready
+sleep infinity &
+exec pgctl-poll-ready sh -c 'sleep 2; touch readyfile; exec sleep infinity'
