+++ tests/examples/orphan-subprocess/playground/slow-startup/timeout-ready
@@ -1,0 +1,1 @@
+5
