+++ tests/spec/examples.py
@@ -17,8 +17,6 @@ from testing.assertions import assert_svstat
 from testing.assertions import wait_for
 
 from pgctl.daemontools import SvStat
-from pgctl.errors import LockHeld
-from pgctl.functions import check_lock
 
 parametrize = pytest.mark.parametrize
 
@@ -599,121 +597,3 @@ pgctl-2015: error: too few arguments
 ''',
                 2,
             )
-
-
-class DirtyTest(object):
-
-    LOCKERROR = '''\
-Stopping: sweet
-ERROR: service sweet timed out at {time} seconds: The supervisor has stopped, but these processes did not:
-UID +PID +PPID +PGID +SID +C +STIME +TTY +STAT +TIME +CMD
-\\S+ +\\d+ +\\d+ +\\d+ +\\d+ +\\d+ +\\S+ +\\S+ +\\S+ +\\S+ +{cmd}
-
-temporary fix: lsof -t playground/sweet | xargs kill -9
-permanent fix: http://pgctl.readthedocs.org/en/latest/user/quickstart.html#writing-playground-services
-
-Stopped: sweet
-ERROR: Some services timed out.
-$'''
-
-    @fixture(autouse=True)
-    def cleanup(self, in_example_dir):
-        try:
-            yield in_example_dir
-        finally:
-            # we use SIGTERM; SIGKILL is cheating.
-            print('killing.')
-            limit = 100
-            while limit > 0:  # pragma: no cover: we don't expect to ever hit the limit
-                try:
-                    check_lock('playground/sweet')
-                    break
-                except LockHeld:
-                    cmd = 'lsof -tau $(whoami) playground/sweet | xargs --replace kill -TERM {}'
-                    Popen(('sh', '-c', cmd)).wait()
-                    limit -= 1
-
-
-class DescribeOrphanSubprocess(DirtyTest):
-
-    @fixture
-    def service_name(self):
-        yield 'orphan-subprocess'
-
-    def it_starts_up_fine(self):
-        assert_command(
-            ('pgctl-2015', 'start'),
-            '',
-            '''\
-Starting: sweet
-Started: sweet
-''',
-            0,
-        )
-        assert_command(
-            ('pgctl-2015', 'log'),
-            '''\
-==> playground/sweet/stdout.log <==
-sweet
-
-==> playground/sweet/stderr.log <==
-sweet_error
-''',
-            '',
-            0,
-        )
-
-    def it_shows_error_on_stop(self):
-        assert_command(
-            ('pgctl-2015', 'start'),
-            '',
-            '''\
-Starting: sweet
-Started: sweet
-''',
-            0,
-        )
-        assert_command(
-            ('pgctl-2015', 'restart'),
-            '',
-            S(self.LOCKERROR.format(time='2', cmd='sleep infinity')),
-            1,
-        )
-
-
-class DescribeSlowShutdown(DirtyTest):
-    """This test case takes three seconds to shut down"""
-
-    @fixture
-    def service_name(self):
-        yield 'slow-shutdown'
-
-    @fixture(autouse=True)
-    def environment(self):
-        os.environ['PGCTL_TIMEOUT'] = '1.5'
-        yield
-        del os.environ['PGCTL_TIMEOUT']
-
-    def it_fails_by_default(self):
-        check_call(('pgctl-2015', 'start'))
-        assert_svstat('playground/sweet', state='up')
-        assert_command(
-            ('pgctl-2015', 'stop'),
-            '',
-            S(self.LOCKERROR.format(time='1\\.5', cmd='sleep 2\\.25')),
-            1,
-        )
-
-    def it_can_shut_down_successfully(self):
-        # if we configure it to wait a bit longer, it works fine
-        with open('playground/sweet/timeout-stop', 'w') as timeout:
-            timeout.write('3')
-
-        check_call(('pgctl-2015', 'start'))
-        assert_svstat('playground/sweet', state='up')
-
-        check_call(('pgctl-2015', 'restart'))
-        assert_svstat('playground/sweet', state='up')
-
-        check_call(('pgctl-2015', 'stop'))
-        assert_svstat('playground/sweet', state=SvStat.UNSUPERVISED)
